Add all files into single 'bob' package
