+++ src/main/java/bob/Command.java
@@ -1,3 +1,4 @@
+package bob;
 public abstract class Command {
     public abstract void execute(Ui ui);
 }
