+++ src/main/java/bob/DeleteCommand.java
@@ -1,23 +1,26 @@
+package bob;
 import java.io.IOException;
 
-public class MarkCommand extends Command {
-    
+public class DeleteCommand extends Command {
+
     private int number;
 
-    public MarkCommand(int number) {
+    public DeleteCommand(int number) {
         this.number = number;
     }
-    
+
     @Override
     public void execute(Ui ui) {
         try {
-            TaskList.markTask(number);
-            ui.print("Bob is on it! Marked the following as done [X]:", 
-                    TaskList.getTask(number).toString());
+            Task task = TaskList.getTask(number);
+            TaskList.deleteTask(number);
+            ui.print("Bob is on it! Deleted this task: ", 
+                    task.toString(), 
+                    "Now you have " + TaskList.getCount() + " task(s).");
         } catch (IndexOutOfBoundsException e) {
             ui.print("Uh oh! Bob says...the task number does not exist.");
         } catch (IOException e) {
             ui.print("Uh oh! Bob says...I couldn't save the task to the file.");
-        }
+        }   
     }
 }
