+++ src/main/java/bob/Event.java
@@ -1,3 +1,4 @@
+package bob;
 import java.time.LocalDateTime;
 
 public class Event extends Task {
