+++ src/main/java/bob/EventCommand.java
@@ -1,19 +1,21 @@
+package bob;
 import java.io.IOException;
+import java.time.LocalDateTime;
 
-public class TodoCommand extends Command {
+public class EventCommand extends Command {
     
-    private Todo todo;
+    private Event event;
 
-    public TodoCommand(String description) {
-        this.todo = new Todo(description);
+    public EventCommand(String description, LocalDateTime from, LocalDateTime to) {
+        this.event = new Event(description, from, to);
     }
     
     @Override
     public void execute(Ui ui) {
         try {
-            TaskList.addTask(this.todo);
+            TaskList.addTask(this.event);
             ui.print("Bob is on it! I've added this task:", 
-                    this.todo.toString(), 
+                    this.event.toString(), 
                     "Now you have " + TaskList.getCount() + " task(s).");
         } catch (IOException e) {
             ui.print("Uh oh! Bob says...I couldn't save the task to the file.");
