+++ src/main/java/bob/ExitCommand.java
@@ -1,3 +1,4 @@
+package bob;
 public class ExitCommand extends Command {
     @Override
     public void execute(Ui ui) {
