+++ src/main/java/bob/ListCommand.java
@@ -1,3 +1,4 @@
+package bob;
 public class ListCommand extends Command {
     @Override
     public void execute(Ui ui) {
