+++ src/main/java/bob/Parser.java
@@ -1,3 +1,4 @@
+package bob;
 import java.io.IOException;
 import java.time.LocalDateTime;
