+++ src/main/java/bob/Storage.java
@@ -1,3 +1,4 @@
+package bob;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
