+++ src/main/java/bob/TodoCommand.java
@@ -1,20 +1,20 @@
+package bob;
 import java.io.IOException;
-import java.time.LocalDateTime;
 
-public class DeadlineCommand extends Command {
+public class TodoCommand extends Command {
     
-    private Deadline deadline;
+    private Todo todo;
 
-    public DeadlineCommand(String description, LocalDateTime by) {
-        this.deadline = new Deadline(description, by);
+    public TodoCommand(String description) {
+        this.todo = new Todo(description);
     }
     
     @Override
     public void execute(Ui ui) {
         try {
-            TaskList.addTask(this.deadline);
+            TaskList.addTask(this.todo);
             ui.print("Bob is on it! I've added this task:", 
-                    this.deadline.toString(), 
+                    this.todo.toString(), 
                     "Now you have " + TaskList.getCount() + " task(s).");
         } catch (IOException e) {
             ui.print("Uh oh! Bob says...I couldn't save the task to the file.");
