+++ src/main/java/bob/Ui.java
@@ -1,3 +1,4 @@
+package bob;
 import java.util.List;
 import java.util.Scanner;
