+++ src/main/java/bob/UnmarkCommand.java
@@ -1,25 +1,24 @@
+package bob;
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
-
+public class UnmarkCommand extends Command {
+    
     private int number;
 
-    public DeleteCommand(int number) {
+    public UnmarkCommand(int number) {
         this.number = number;
     }
-
+    
     @Override
     public void execute(Ui ui) {
         try {
-            Task task = TaskList.getTask(number);
-            TaskList.deleteTask(number);
-            ui.print("Bob is on it! Deleted this task: ", 
-                    task.toString(), 
-                    "Now you have " + TaskList.getCount() + " task(s).");
+            TaskList.unmarkTask(number);
+            ui.print("Bob is on it! Marked the following as undone [ ]", 
+                    TaskList.getTask(number).toString());
         } catch (IndexOutOfBoundsException e) {
             ui.print("Uh oh! Bob says...the task number does not exist.");
         } catch (IOException e) {
             ui.print("Uh oh! Bob says...I couldn't save the task to the file.");
-        }   
+        }
     }
 }
