+++ src/main/java/bob/WrongCommandException.java
@@ -1,3 +1,4 @@
+package bob;
 
 
 public class WrongCommandException extends Exception {
